add react hook feat
